+++ steps/pluginsonresponse.go
@@ -28,6 +28,7 @@ import (
 
 	"github.com/Sirupsen/logrus"
 	"github.com/northwesternmutual/kanali/controller"
+	"github.com/northwesternmutual/kanali/metrics"
 	"github.com/northwesternmutual/kanali/plugins"
 	"github.com/northwesternmutual/kanali/spec"
 	"github.com/northwesternmutual/kanali/utils"
@@ -44,7 +45,7 @@ func (step PluginsOnResponseStep) GetName() string {
 }
 
 // Do executes the logic of the PluginsOnResponseStep step
-func (step PluginsOnResponseStep) Do(ctx context.Context, c *controller.Controller, w http.ResponseWriter, r *http.Request, resp *http.Response, trace opentracing.Span) error {
+func (step PluginsOnResponseStep) Do(ctx context.Context, m *metrics.Metrics, c *controller.Controller, w http.ResponseWriter, r *http.Request, resp *http.Response, trace opentracing.Span) error {
 	untypedProxy, err := spec.ProxyStore.Get(r.URL.Path)
 	if err != nil || untypedProxy == nil {
 		return utils.StatusError{Code: http.StatusNotFound, Err: errors.New("proxy not found")}
@@ -60,7 +61,7 @@ func (step PluginsOnResponseStep) Do(ctx context.Context, c *controller.Controll
 		if err != nil {
 			return err
 		}
-		if err := doOnResponse(ctx, plugin.Name, proxy, *c, r, resp, trace, *p); err != nil {
+		if err := doOnResponse(ctx, m, plugin.Name, proxy, *c, r, resp, trace, *p); err != nil {
 			return err
 		}
 	}
@@ -68,7 +69,7 @@ func (step PluginsOnResponseStep) Do(ctx context.Context, c *controller.Controll
 	return nil
 }
 
-func doOnResponse(ctx context.Context, name string, proxy spec.APIProxy, ctlr controller.Controller, req *http.Request, resp *http.Response, span opentracing.Span, p plugins.Plugin) (e error) {
+func doOnResponse(ctx context.Context, m *metrics.Metrics, name string, proxy spec.APIProxy, ctlr controller.Controller, req *http.Request, resp *http.Response, span opentracing.Span, p plugins.Plugin) (e error) {
 	defer func() {
 		if r := recover(); r != nil {
 			logrus.Errorf("OnResponse paniced: %v", r)
@@ -79,7 +80,7 @@ func doOnResponse(ctx context.Context, name string, proxy spec.APIProxy, ctlr co
 	sp := opentracing.StartSpan(fmt.Sprintf("PLUGIN: ON_RESPONSE: %s", name), opentracing.ChildOf(span.Context()))
 	defer sp.Finish()
 
-	if err := p.OnResponse(ctx, proxy, ctlr, req, resp, sp); err != nil {
+	if err := p.OnResponse(ctx, m, proxy, ctlr, req, resp, sp); err != nil {
 		return err
 	}
 	return nil