+++ steps/writeresponse_test.go
@@ -28,6 +28,7 @@ import (
 	"net/http/httptest"
 	"testing"
 
+	"github.com/northwesternmutual/kanali/metrics"
 	opentracing "github.com/opentracing/opentracing-go"
 	"github.com/stretchr/testify/assert"
 )
@@ -48,7 +49,7 @@ func TestWriteResponseDo(t *testing.T) {
 		},
 		Body: bytes.NewBuffer([]byte("this is my mock response body")),
 	}
-	err := step.Do(context.Background(), nil, writer, nil, response.Result(), opentracing.StartSpan("test span"))
+	err := step.Do(context.Background(), &metrics.Metrics{}, nil, writer, nil, response.Result(), opentracing.StartSpan("test span"))
 	defer writer.Result().Body.Close()
 	assert.Nil(t, err)
 	assert.Equal(t, writer.Result().StatusCode, 200)